generator/code/requirements.txt
mysql-connector-python==8.0.25
protobuf==3.17.3